Separate Formulas Into New Util
